Add files via upload
PREREQUISITES- phpMyAdmin/XAMPP-extract the folder in the following path C:\xampp\htdocs.
All the images used are compressed in the folder named ResourcesNEW therefore they will have to be extracted prior usage of the site. the.SQL file in the folder SQL DATABASE will have to be imported in your phpMyAdmin as localhost.